added simple sketch of readonly public members
This works so far but has the drawback, that all
private and protected members are effected to and
exposed as public.

This could easily be solved by reflection, but will cost
more performance I think. In the scope of this project
we will create a lot of this data objects, which will
have no private members.

So for now it's ok, but ideas on how to solve this, without
reflection are welcome!